Move the 'Command' class from 'applications' to its own 'command' module.
